feat (embed): embed player on friendly site
implements https://trello.com/c/yDdzf9cL/73-embeddability

see example in `/examples/embed/friendly.html`
