+++ src/examples/embed/friendly.html
@@ -8,6 +8,7 @@
       src="/js/podlove-web-moderator.js"
       data-podlove-web-player-source="/examples/cre-206/"&gt;&lt;/script&gt;
   </code></pre>
-  <script src="/js/podlove-web-moderator.js" data-podlove-web-player-source="/examples/cre-206/"></script>
+  <script src="/js/podlove-web-embed.js" rel="embed"
+    data-podlove-web-player-source="/examples/cre-206/"></script>
 </body>
 </html>
